+++ traffic_api/views.py
@@ -167,25 +167,3 @@ def historical_request(request):
 
     # Only need to aggregate over timestamps if not hourly request.
     return JsonResponse(response, safe=False)
-
-
-def floor_date(ts, granularity):
-    '''
-    Can pass in one of the granularities, floors the timestamp to the given
-    granularity.
-
-    '''
-
-    dt = datetime.datetime.fromtimestamp(ts)
-
-    if granularity == 'yearly':
-        dt_floor = dt.replace(month=1, day=1, hour=0, minute=0, second=0)
-    elif granularity == 'monthly':
-        dt_floor = dt.replace(day=1, hour=0, minute=0, second=0)
-    elif granularity == 'daily':
-        dt_floor = dt.replace(hour=0, minute=0, second=0)
-    else:
-        dt_floor = dt
-
-    # Convert back to timestamp and add 8 hours (UTC).
-    return int(time.mktime(dt_floor.timetuple()) + 8*3600)
